implement htmlContent setter and getter on the RawEditorController
